Implemented conditional accessibility extensions. New option.
Now accessibility related behaviour of AccessibleToast,
ExplorableTextView, AccessibleLayout, AccessibleActivity and
AccessibleTrackedActivity is optional. It is on when both global
accessibility means in system and accessibility extensions in OsmAnd
are turned on. Otherwise all these classes behave exactly identical to
their direct parents, so, hopefully, can be safely used in place of
them in all contexts.