bicycle routing: Don't prefer cycleways too much over tracks/path/footways
Previously we would make huge detours to prefer cycleways. E.g. versus a
footway with bicycle=yes we would take more than 3 time as long ways
(16/8)*(1.5/0.9) > 3 to use a cycleway. We would also take long detours
(more than twice as long ways) to prefer a track/path over a footway with
bicycle=yes. Often footway/bicycle=yes are synonyms for highway=path,
and the latter is not faster or better at all. This led to weird bicycle
routing where I was taking long detours to stay on a "path" rather than
a "footway".

Take up Victors suggestion of bumbing the speed of a footway
from 8 to 12, which is still realistic and tone down the priority of 1.5
for cycleways/track/path to 1.3 which was too high.
This is what this patch does. Now we have:

Highway=   Speed=   Priority=  Detours-vs-footway=
footway    12       1          --
track/path 12       1.3        30%
cycleway   16       1.3        73%

I believe this still prefers tracks/path and cycleway enough and the
weird routing went away. (the data has changed now or I would give
the example location. But I tested a few bike routes involving various
ways and the suggestions were sensible.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>