Fix inefficient layout weight
When only a single widget in a LinearLayout defines a weight, it is more
efficient to assign a width/height of 0dp to it since it will absorb all
the remaining space anyway. With a declared width/height of 0dp it does
not have to measure its own size first.