New plugin for OsmAnd to provide separate display magnification controls as well as map a zoom level selector.
The plugin allows the additional magnification buttons to be made visible or hidden with an override of the long press dialog on the map zoom buttons.

With the new buttons displayed, a long press on the map zoom buttons will bring up a selector dialog to choose any level to be displayed.
The magnification selection dialog is activated by a long press of either of the magnifications buttons.

Only minor changes to the existing code base are required to implement this plugin.

This pull request is independent of any other pull request associated with the Distance Calculator plugin.