Implement merging nearby turns together.
When there are multiple turns near each other (< 60m), set the lanes
that are allowed on earlier turns such that staying on that lane will
carry you through the other turns.

For a standard left/right turn at a traffic light at two one-way roads
(dual carriageway), this means that the left and right turns are
correctly highlighted.