if incoming URL was properly parsed, remove GeoIntentActivity from backstack
When a user clicks on a URL that OsmAnd claims, they will first go to the
GeoIntentActivity, which will show them the processing, and any errors,
then it'll take them to the map.  Since the map will stay in the same task
as the original Activity where the user clicked it, the user is very likely
to be clicking on the back button.  Calling finish() here gives them one
less thing to click back on, since GeoIntentActivity will be removed.

closes #1037 https://github.com/osmandapp/Osmand/pull/1037