claim all Google map domains
This is all of the possible Google Maps domains.  Unfortunately for this
case, Google redirects the .com to the country-specific domain of the
detection location of the user.  So if a user copies and pastes a map URL,
then it'll include that user's country domain.  So in order to support
catching Google Maps links well, the country domains need to be supported.

https://en.wikipedia.org/wiki/List_of_Google_domains