tighten parseLength() again
I wonder what the test for "-" does. There can't be any negative values in maxlength or maxheight.
It seems your version also accepts values like "1 km 1 km" (meaning 2km) which is invalid per the wiki, but probably shouldn't harm. I'm fine with that. But I think it also accepted "1 mile" and thinking it was 1meter. You can't just test unit.contains("m") as that would also match "mile", "cm", etc. That is not good so I tightened it a bit.

So I haven't compiled this again, please tell me if all the tests pass.