Reworked GPX display.
1. GPX track colour can come from .gpx file (as before) via <extensions> block. However, this was previously
 on an entire-track basis. Since the reader code also worked when an <extensions> block was placed inside a <trk> or
<trkseg> block, I have honoured this implicit arrangement.  So, if an extension containing <color> is inside a <trkseg>
then that segment will be drawn in the selected colour. If there is no colour specified for a segment, it will draw
in it's parent track's colour. If the parent's colour is not specified, it will draw in the GPX file's colour. If
none is specified then the default OsmAnd colour is used.

2. Track width is also configurable on per-segment. I added <zoom>1.0</zoom> into the <extensions>, so if you put
that in either a gpx <trk> or <trkseg> then you will get scaling of the width of the drawn segments. The zoom is
relative to default size, so 1.0 is 100%. To draw double-thick lines, use 2.0, or half-width 0.5.

With (1) and (2) you can use one or both inside the <extensions> block.

The reasoning for the above is that you may wish to have some tracks marked with easy and difficult sections, or
by different surface types, etc., By putting colouring capability in segments you can do all of that.

3) Reduced points for speedups.  Huge GPX files with hundreds of thousands of points could potentially slow the UI down.
I have implemented point-reduction at runtime so that tracks are displayed using the minimum number of points required
to maintain visual integrity. Code is inside GPXLayer.java and uses Ramer-Douglas-Peucer algorithm. Selection criteria
for culling points is based on view zoom level.  The culled point set is cached, so the line reduction only happens when
there is a change in zoom, or number of points in the original line. The reduction could theoretically replace the
"offscreen" culling code, because that code has a number of bugs;

  - e.g., 1) lines just offscreen are clipped but you can see curved ends on them when they shouldn't be.
  2) some lines that should be onsceen are totally culled. I have a sample track demonstrating this bug.

My preference would be to simply rely on the point culling to give great perfomance and drop the segment onscreen checks.

4) Scaling based on view zoom - the line widths are modified according to the view zoom modified by (2) the track
zoom. Now when you zoom right out, you don't get a large block of "mess" - in fact tracks are legible right down to very
small sizes (zooming right out).  It's a lovely thing.

Andrew Davie
andrew@taswegian.com