How to please everyone!
This code has a switch - autoScale.
It defaults to false (off).

Behavoiur when off - as requested.
That is, ONLY segments from GPX tracks which have specific zoom set will be auto-scaled.
By auto-scaled, I mean the track width changes as you zoom in and out in the UI.

The other option true (on)...
There is an access function if you want to turn on the autoscaling for everything.

Behavior when on - all things are auto-scaled.
This works if you have just a single track width at all scales (as the original code did).

Why doesn't it seem to work when I change the track width in the render XML being used?
--> because the original code in updatePaints uses a hash to avoid running more than once, and the track width
 is NOT a part of that hash. If the updatePaints code that initialises the track width doesn't run, then my
 code is never "told" that there's a change.  So, fix that hash :)

Summary:

* Default behaviour - only customised GPX tracks with embedded zooms will automatically scale
* GPX widths set via render options are left unchanged.
* Optional switch to use the automatic zoom for everything (in the case where render options are absent and just a single track width is specified).