I've added this could-be-useful routine called 'resampleTrack'. What it does is take a list of points, and resamples it to a new list with points a set distance apart. The new points are always ON the original track; they are just evenly spaced.
This could be useful for placing arrow/markers on a track, for example - getting the spacing right.
It gives you accurate distance points. Could be used for displaying coloured segments.

For example, to give a evenly-spaced colouring of a track based on altitude, call this routine on the original track, and
use the evenly-spaced segments to colour things accuractely.

Hope this might be useful.  Could probably be moved to a more suitable file like GPXUtilities.java

I have tested this - you can see the test code still in there where the culledPoints list is generated. I'm displaying the
resampled track instead of the culled track, and it works just fine.