Especially for Victor!
The RamerDouglasPeucer track reduction now operates asynchronously. Implemented via AsyncTask.

Here's the process;

If a zoom resize or track change is detected, then
 - cancel any existing cull operations still in progress for this track
 - start a new cull process for this track
 - set this track to full resolution for display purposes

In the asynchronous thread:
 - do the culling
 - if finished and not cancelled (see above) then tell the UI the culled track is available
 - force a UI repaint (via view.refreshMap() - is this correct?)

So now it's a lovely thing - no delay at all when zooming and re-culling the tracks.
When a culled track is not available (i.e., initially or when a cull is in operation and the existing cull track is
inappropriate) then the full-resolution track is used for the drawing instead.

Seems to work beautifully.