Arrows renderer now working reasonably well.
I put a bit of effort into "pretty". Also, arrows are not displayed until a particular zoom level.
Sample video of this version with the following renderers active: rainbow altitude band, original track, distance
markers and arrows. See video at https://youtu.be/FNui0zxJGJI

Seems to be pretty solid for all renderers. Speed is good. This is close to a "final" version.