Bugfix: Current Track disappeared when starting point wandered offscreen
The 'bounds' that I had implemeted earlier described the minimal rectangle inside which a whole track was contained. This
bounding rectangle was used to quickly cull all track types at draw-time. And it worked well... too well!  The current track
is continually adding points to itself, so its bounds need to be recalculated regularly, not just at startup like other
track types.

So, I've implemented a new Renderable, 'CurrentTrack' which does the special handling for that type of track. It is
smart in that it only recalculates/updates bounding rectangle on the newly added points, so it's very quick. I tested
it on my phone and it appears to be working OK.

Other minor adjustments here and there as I tweak things.