Added speed and altitude renderers
Removed currentTrack renderer - this is now handled by the StandardTrack renderer.
Renderer instantiations now have 1st parameter "priority". Renderers are sorted before display
so that all (for example) altitude colour bands can be rendered before all actual tracks - no
overlap/underlap problems.

Speed and altitude are resampled in the same way that tracks are; that is, they are asynchronous
and reduce point count to keep the shape of tracks the same with minimal points.

There are some variations in the colours as you zoom in/out (in localised areas) because there
are simply fewer points and the colours of the segments are end-point colours not averaged over
the entire segment. It's pretty good though.

Note that these colour bands range from purple (low) to red (high) and that the range represents
the variation across each track. They are not absolute values. That is, red does not represent
1000 m high - or whatever. It simply indicates the highest altitude on that particular track.