ltitude/Speed colour banding, distance marking,  and efficiency improvements
To instantiate each type of renderer...

[Current Track]
ts.renders.add(new Renderable.CurrentTrack(ts.points, 0));

This type of renderer is specific to the 'current track' recording a GPX track. It checks for additional points being
added to the original track, and gracefully handles those. It appends any new points to the 'culled' track (the one
that has been resampled to minimise size, based on scale) and when it gets a chance, it re-culls the track to incorporate
the newly added points. Completely transparent to the user - it should just look right.

Note: probably best not to add altitude/speed/distance to the 'current track'.

[Standard Track]
s.renders.add(new Renderable.StandardTrack(ts.points, epsilon));

Draws a stock-standard GPX track in whatever colour is already set in the paint context. The 'epsilon' parameter
controls how far off a true-line between two points A, B, a third point C must be before it is culled as a part of the
line reduction algorighm. The value is a zoom value and is used to calculate an actual metre deviation orthogonal
distance minimum. Increase this value to allow more distance and less accurate lines  (but smaller number of points).
At 16.5, there is no noticeable difference between original and culled tracks - even though the culled tracks are
significantly smaller.

[Altitude]
ts.renders.add(new Renderable.Altitude(ts.points, epsilon, 2.5));

The epsilon parameter is the same as per Standard Track. The last parameter is a width multiple of the track size.
A colour-band is placed behind the Standard track, where the colour is based on a graduated rainbow scale (purple is low
and red is high). The colouration comes from analysing altitude (ele) variation along the entire track - so note:
this is not an ABSOLUTE altitude, it is an altitude relative indicator for the track itself.

[Speed]
ts.renders.add(new Renderable.Speed(ts.points, epsilon, 4));

Same as Altitude, except it uses calculated speed to drive the colour banding. The priority of the speed band is
LOWER than the altitude band, so if both are enabled for a track segment, then the speed will be drawn first, then
the altitude, then of course the track itself.  Note that the priorities are hardwired in the code.

Note, too, that ALL speed bands are drawn first, then ALL altitude bands, then ALL tracks - in other words, speed band
always always goes under tracks.

[Distance]
ts.renders.add(new Renderable.Distance(ts.points, Renderable.Distance.unit.KILOMETERS));

Draws text and a marker on a track at given distance spacing. The text will not be displayed until a particular zoom
level, based on the spacing of the distances. These are configured in "public enum unit" in the Distance class. Note
that there is inbuilt support for METERS, METRES, KILOMETERS, KILOMETRES, MILES, YARDS, FEET - just change the
enum used as the last parameter. The correct abbreviation is included with the text output.