Added (optional) speed and altitude colour band range setting.
Speed and altitude colour bands auto-calculate the minimum and maximum values along a track, and then colour each part
of the track according to this range.  This gives excellent rainbow colour for any track, as you can immediately see
the highest and lowest parts (or slowest/fastest) but each track is independent.

Now you can call .setRange(min,max) on speed/altitude renderers to set the extremes of the range for the colour banding.
For example...

	Renderable.Altitude alt = new Renderable.Altitude(view, ts.points, epsilon, 2.5);
	alt.setRange(0,500);	// optional  - set colour range 0m to 500m

The setRange call is optional. Here, as it is present, it causes the colour banding calculation to clamp the range
to between 0 and 500 (metres). Any values outside that range are clamped to the relevant high/low value.  The value
of this way of doing things is that multiple tracks will have consistent colouring - 500m is the same colour no matter
what track is displayed.

So, best of both worlds - original behaviour if you want to see the range over the track and have maximum use of colours
to differentiate the variations, or clamped behaviour if you want to have consistent colouring across tracks.

It seems to me that it would be useful to have a colour scale displayed on the screen, with the various graduations and
values for colours shown.  For example, a vertical band on the side of the screen with numbers overlayed.