Fix #2484 - GPX tracks not shown unless extension is .gpx
Before this change, GPX tracks are not shown in "My Tracks" in "My Places"
unless they have a lowercase .gpx extension.

Such tracks are still shown in "Configure map", "GPX track..." but the
extension isn't removed like it is for tracks with a lowercase extension.
This has been fixed too.