Fix loading of custom rendering files containing dashes and underscores
Selecting a custom rendering style whose filename contains one or more
dashes or underscores failed with a "Exception occurred: renderer was not
loaded" message.

This was because since commit 7fc7b5aa373c8fdd8d65429de4c168a2b2bd4f59,
dashes and underscores in custom rendering style names are replaced by
spaces for display and selection, but stored with dashes and underscores in
externalRenderers in RendererRegistry.

The fix is to apply the same processing to each renderer name when
inserting into externalRenderers as when creating visibleNamesList.