Optimize png images to reduce size
To reduce the size of the app I optimized all png images using `zopflipng -m` the opzimization saves just some bytes per png image. Depending on the image it could be more than 90% reduction of size, but most of the time it is less. As zopflipng does lossless reduction it should just work fine, but this should be checked as some devices might have problems with 2-channel gray+alpha png images.