Add an option for always adding track waypoints to the current track
When recording a track (and displaying it), and displaying one or more
tracks at the same time, OsmAnd asks which track to add any track waypoints
to.

This can become tedious when adding many track waypoints (like I do when
using OsmAnd to collect addresses). There's a possibility that you might
tap the wrong track by mistake, particularly if your phone screen is wet or
you're wearing gloves.

This change introduces a new setting for the trip recording plugin. When
enabled (it's disabled by default), track waypoints will always be added to
the current track (OsmAnd will not prompt ask which track to use).