Allow to completely disable map rotation using gesture.
The map rotation using gesture was possible even when "North is up" is set as Map orientation. So, create a separate independent setting to disallow the gesture. Actually, the setting can be useful even in other Map orientation modes.

By default, the gesture is enabled, so the app behaves as before.

Disabling the rotation gesture allows users to avoid accidental rotation when zooming was intended. The problem manifests itself e.g. in these scenarios:
- (1) I zoom the map. The map rotates while zooming. I need to click the compass icon to revert to North is up. Actually I needed to build the habit of "checking if north is up" after each zooming, which is annoying.
- (2) I zoom the map. The map rotates while zooming. I do not notice this. Later, I notice "something is wrong with the map". Click the compass icon.
- (3) During some activity (geocaching, bike, sport) I accidentally rotate the map without the intention to do anything. Usually I do not notice this and it continues as in (2).

This supports the issue 15660 and some related issues.